Adopt ruff; format the codebase and fix what it found
Dependencies were already at their latest versions (uv lock --upgrade was a
no-op), so "latest stack" here meant tooling: there was no linter or
formatter. ruff is the current standard and one dependency.

Configured for correctness and modernisation rather than style: pyflakes,
pycodestyle errors, isort, pyupgrade, bugbear, comprehensions, bandit's
security checks, simplify, and ruff's own rules. Line length is left to the
formatter.

Of 35 findings, 26 were mechanical (quoted annotations, deprecated typing
imports, unused imports and noqa). The rest: two zip() calls now say
whether lengths must match, three redundant int(round()) casts, two unused
unpacked variables, one nested if, one test regex that wanted a raw string.
No behaviour changes; 128 tests pass before and after.

+++ gpp/compile.py
@@ -263,17 +263,13 @@ def content_tag(plan_name: str, payload: dict[str, Any]) -> str:
     editing a step changes the hash, so the push layer knows to replace it.
     The plan slug lets you keep several plans on one calendar.
     """
-    skeleton = json.dumps(
-        payload.get("workoutSegments", []), sort_keys=True, separators=(",", ":")
-    )
+    skeleton = json.dumps(payload.get("workoutSegments", []), sort_keys=True, separators=(",", ":"))
     digest = hashlib.sha256(skeleton.encode("utf-8")).hexdigest()[:8]
     slug = "".join(c for c in plan_name.lower() if c.isalnum())[:16] or "plan"
     return f"[{TAG_PREFIX}:{slug}:{digest}]"
 
 
-def compile_workout(
-    workout: Workout, profile: Profile, plan_name: str = "plan"
-) -> CompiledWorkout:
+def compile_workout(workout: Workout, profile: Profile, plan_name: str = "plan") -> CompiledWorkout:
     sport = SPORT_TYPES.get(workout.sport)
     if sport is None:
         raise CompileError(f"unsupported sport {workout.sport!r}")
@@ -289,7 +285,7 @@ def compile_workout(
     payload: dict[str, Any] = {
         "sportType": sport,
         "workoutName": workout.name,
-        "estimatedDurationInSecs": int(round(estimated)),
+        "estimatedDurationInSecs": round(estimated),
         "workoutSegments": [
             {
                 "segmentOrder": 1,
